Rename PreviewWindow to OutputWindow
