+++ Neo/Parcel.Neo/Helpers/ImageSourceHelper.cs
@@ -2,6 +2,7 @@
 using System.Windows.Media.Imaging;
 using System.Windows.Media;
 using System.IO;
+using Parcel.Neo.PopupWindows;
 
 namespace Parcel.Neo.Helpers
 {
