+++ Neo/Parcel.Neo/PopupWindows/CreateFunctionWindow.xaml
@@ -1,4 +1,4 @@
-﻿<base:BaseWindow x:Class="Parcel.Neo.PreviewWindows.CreateFunctionWindow"
+﻿<base:BaseWindow x:Class="Parcel.Neo.PopupWindows.CreateFunctionWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
