+++ Neo/Parcel.Neo/PopupWindows/LiveCodePreviewWindow.xaml
@@ -1,9 +1,9 @@
-﻿<base:BaseWindow x:Class="Parcel.Neo.PreviewWindows.LiveCodePreviewWindow"
+﻿<base:BaseWindow x:Class="Parcel.Neo.PopupWindows.LiveCodePreviewWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Parcel.Neo.PreviewWindows"
+        xmlns:local="clr-namespace:Parcel.Neo.PopupWindows"
         xmlns:base="clr-namespace:Parcel.Neo"
         mc:Ignorable="d"
         Title="Live Code Preview (Experimental)" SizeToContent="WidthAndHeight" WindowStartupLocation="CenterOwner" MinWidth="400" MinHeight="300"
