+++ Neo/Parcel.Neo/PopupWindows/OutputWindow.xaml
@@ -1,4 +1,4 @@
-﻿<base:BaseWindow x:Class="Parcel.Neo.PreviewWindow"
+﻿<base:BaseWindow x:Class="Parcel.Neo.PopupWindows.OutputWindow"
             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
